partition_tools: extend lpunpack for split partitions
This is meant for retrofitted dynamic partitions where
several super partitions are used. Until now lpunpack
fails with the message

"Split super devices are not supported."

in this case. Extend lpunpack to support zero or more
'-i' arguments with which additional super images may be referenced.

Before running this new tool, do lpdump on the main super image file
and examine the section 'Super partition layout' to get
an idea which additional image files may be needed.

Change-Id: I67904d6cb2a630d1ebe3903db39d23357603a60b